Move crypto test back to test package
