+++ test/tar_ball_test.go
@@ -5,7 +5,6 @@ import (
 	"bytes"
 	"github.com/stretchr/testify/assert"
 	"github.com/wal-g/wal-g/internal"
-	"github.com/wal-g/wal-g/internal/crypter/openpgp"
 	"github.com/wal-g/wal-g/testtools"
 	"io"
 	"strings"
@@ -49,7 +48,7 @@ func TestS3DependentFunctions(t *testing.T) {
 
 	bundle.NewTarBall(false)
 	tarBall := bundle.TarBall
-	tarBall.SetUp(openpgp.MockArmedCrypter())
+	tarBall.SetUp(MockArmedCrypter())
 	tarWriter := tarBall.TarWriter()
 
 	mockData := []byte("a")
@@ -77,7 +76,7 @@ func TestS3DependentFunctions(t *testing.T) {
 	// Test naming property of SetUp().
 	bundle.NewTarBall(false)
 	tarBall = bundle.TarBall
-	tarBall.SetUp(openpgp.MockArmedCrypter(), "mockTarball")
+	tarBall.SetUp(MockArmedCrypter(), "mockTarball")
 	tarBall.CloseTar()
 }
